add filter component, sort function
