+++ src/components/Catalog/catalog.module.scss
@@ -6,30 +6,6 @@
   justify-content: center;
 }
 
-.container {
-}
-
-.filter {
-  display: flex;
-  justify-content: space-between;
-
-  @include on-tablet {
-    justify-content: start;
-    gap: 16px;
-  }
-}
-.filterContainer {
-  display: flex;
-  flex-direction: column;
-  gap: 4px;
-}
-.filterDescription {
-  color: $secondary;
-  font-weight: 500;
-  font-size: 12px;
-  line-height: 15.34px;
-}
-
 .sectionWrapper {
   @include pageGrid;
   margin: 24px 0;
