+++ src/components/UI/ButtonSlider/ButtonSlider.tsx
@@ -1,5 +1,4 @@
 import React from 'react';
-import { Link } from 'react-router-dom';
 import styles from './buttonSlider.module.scss';
 
 interface ButtonSliderProps {
@@ -8,16 +7,19 @@ interface ButtonSliderProps {
   handleClick?: () => void;
 }
 
-export const ButtonSlider: React.FC<ButtonSliderProps> = ({ iconType, active, handleClick }) => {
+export const ButtonSlider: React.FC<ButtonSliderProps> = ({
+  iconType,
+  active,
+  handleClick,
+}) => {
   return (
-    <Link
-      to="#"
+    <div
       className={`${styles.button} ${active ? styles.disable : styles.default}`}
       onClick={handleClick}
     >
       <div
         className={`${styles.icon} ${active ? styles[iconType] : `${styles[iconType]} ${styles.disable}`}`}
       />
-    </Link>
+    </div>
   );
 };
