+++ src/pages/PhonesPage/PhonesPage.tsx
@@ -1,39 +1,55 @@
-import styles from './phonesPage.module.scss';
+import React, { useEffect, useState } from 'react';
 import { useSelector } from 'react-redux';
+import { useSearchParams } from 'react-router-dom';
 import { RootState } from '../../store/store';
+import { Category } from '../../types/Category';
 import { Product } from '../../types/Product';
 import { Breadcrumbs } from '../../components/Breadcrumbs';
 import { Catalog } from '../../components/Catalog';
-import { SetStateAction, useState } from 'react';
 import { Pagination } from '../../components/Pagination';
-import { Category } from '../../types/Category';
+import { sortProducts } from '../../utils/sortProducts';
+
+import styles from './phonesPage.module.scss';
+import { SortBy } from '../../components/Filter';
 
 export const PhonesPage = () => {
-  let products = useSelector((state: RootState) => state.product.products);
+  const products = useSelector((state: RootState) => state.product.products);
   const productsPerPage = useSelector((state: RootState) => state.product.productsPerPage);
   const [currentPage, setCurrentPage] = useState(1);
+  const [sortBy, setSortBy] = useState<SortBy>(SortBy.age);
 
-  products = products.filter(
-    (product: Product) => product.category === Category.phones,
+  const filteredProducts = products.filter(
+    (product: Product) => product.category === Category.phones
   );
 
-  const handlePagination = (pageNumber: SetStateAction<number>) => {
+  const [searchParams] = useSearchParams();
+  useEffect(() => {
+    const params = new URLSearchParams(searchParams);
+    const sortByParam = params.get('sort');
+    if (sortByParam && Object.values(SortBy).includes(sortByParam as SortBy)) {
+      setSortBy(sortByParam as SortBy);
+    } else {
+      setSortBy(SortBy.age);
+    }
+  }, [searchParams]);
+
+  const sortedProducts = sortProducts(filteredProducts, sortBy);
+
+  const handlePagination = (pageNumber: number) => {
     setCurrentPage(pageNumber);
   };
 
   const indexOfLastPost = currentPage * productsPerPage;
   const indexOfFirstPost = indexOfLastPost - productsPerPage;
-  const currentProducts = products.slice(indexOfFirstPost, indexOfLastPost);
+  const currentProducts = sortedProducts.slice(indexOfFirstPost, indexOfLastPost);
 
   return (
     <div className={styles.container}>
       <Breadcrumbs />
-
       <h1 className="title">Mobile phones</h1>
-
-      <Catalog products={currentProducts} totalProducts={products.length}/>
+      <Catalog products={currentProducts} totalProducts={sortedProducts.length} />
       <Pagination
-        length={products.length}
+        length={sortedProducts.length}
         currentPage={currentPage}
         handlePagination={handlePagination}
       />
